+++ src/main/java/com/example/demo/dto/VoucherDto.java
@@ -6,10 +6,10 @@ import java.util.UUID;
 public class VoucherDto {
 
     private final UUID id;
-    private final Integer discountAmount;
+    private final double discountAmount;
     private final VoucherType voucherType;
 
-    public VoucherDto(UUID id, int discountAmount, VoucherType voucherType) {
+    public VoucherDto(UUID id, double discountAmount, VoucherType voucherType) {
         this.id = id;
         this.discountAmount = discountAmount;
         this.voucherType = voucherType;
@@ -22,8 +22,8 @@ public class VoucherDto {
     @Override
     public String toString() {
         return switch (voucherType) {
-            case FIXED_AMOUNT_VOUCHER -> String.format("    Fixed Voucher, Discount Amount: %d 할인권이 생성되었습니다.", discountAmount);
-            case PERCENT_DISCOUNT_VOUCHER -> String.format("    Percent Voucher, Discount percent Amount: %d 할인권이 생성되었습니다.", discountAmount);
+            case FIXED_AMOUNT_VOUCHER -> String.format("    Fixed Voucher, Discount Amount: %.0f 할인권이 생성되었습니다.", discountAmount);
+            case PERCENT_DISCOUNT_VOUCHER -> String.format("    Percent Voucher, Discount percent Amount: %.0f 할인권이 생성되었습니다.", discountAmount);
             default -> throw new IllegalArgumentException("잘 못된 바우처 타입입니다.");
         };
     }
